Added optional countdown hack. Edit bwapi.ini to use it. Attempt to fix several inconsistencies and warnings. Can't test since it doesn't compile anyway.
git-svn-id: https://bwapi.googlecode.com/svn/trunk@1790 273ecf2c-f24b-0410-b6a7-e9298bb4873f

+++ bwapi/BWAPI/Source/BW/UnitType.h
@@ -50,9 +50,9 @@ namespace BW
       u8                                   requiredTech() const;
       u8                                   armorUpgrade() const;
 
-      u16                  maxHitPoints() const;
-      u16                  maxShields() const;
-      u16                  maxEnergy() const;
+      s32                  maxHitPoints() const;
+      s16                  maxShields() const;
+      u8                   maxEnergy() const;
       u8                   armor() const;
 
       u16                  mineralPrice() const;
